requirements.txt
# ============================================================
# Automated Book Generation System — Python Dependencies
# ============================================================

# --- AI / LLM ---
google-genai>=0.1.0

# --- Database ---
supabase>=2.4.0

# --- Excel Input ---
openpyxl>=3.1.2

# --- DOCX Output ---
python-docx>=1.1.0

# --- PDF Output ---
reportlab>=4.1.0

# --- HTTP Client  ---
httpx>=0.27.0

# --- Environment / Config ---
python-dotenv>=1.0.0

# --- Utilities ---
tenacity>=8.3.0        # retry logic for LLM/API calls
pydantic>=2.6.0        # data validation (supplementary to dataclasses)

# --- Dev / Testing ---
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
